complete

+++ hapus_subjek.php
@@ -25,22 +25,22 @@ $delTop = $hubung->prepare("DELETE FROM topik WHERE idsubjek = ?");
 $delTop->bind_param('i', $del_subjek);
 $delTop->execute();
 
-#if (isset($infoDel['idsoalan'])){
-$delPil = $hubung->prepare("DELETE FROM pilihan WHERE idsoalan = ?");
-$delPil->bind_param('i', $infoDel['idsoalan']);
-$delPil->execute();
-#}
+if (isset($infoDel['idsoalan'])){
+	$delPil = $hubung->prepare("DELETE FROM pilihan WHERE idsoalan = ?");
+	$delPil->bind_param('i', $infoDel['idsoalan']);
+	$delPil->execute();
+}
 
 
-#if (isset($infoDel['idtopik'])){
-$delRek = $hubung->prepare("DELETE FROM perekodan WHERE idtopik = ?");
-$delRek->bind_param('i', $infoDel['idtopik']);
-$delRek->execute();
+if (isset($infoDel['idtopik'])){
+	$delRek = $hubung->prepare("DELETE FROM perekodan WHERE idtopik = ?");
+	$delRek->bind_param('i', $infoDel['idtopik']);
+	$delRek->execute();
 
-$delSoal = $hubung->prepare("DELETE FROM soalan WHERE idtopik = ?");
-$delSoal->bind_param('i', $infoDel['idtopik']);
-$delSoal->execute();
-#}
+	$delSoal = $hubung->prepare("DELETE FROM soalan WHERE idtopik = ?");
+	$delSoal->bind_param('i', $infoDel['idtopik']);
+	$delSoal->execute();
+}
 
  //Papar mesej jika berjaya hapus
 echo "<script>alert('Hapus Subjek Berjaya!'); window.location='subjek_senarai.php'</script>";
